requirements.txt
# requirements.txt
Flask
pandas
plotly
openpyxl
psycopg2-binary
gunicorn
